rename chemistry.featurizer to chemistry.fingerprint
