+++ glasspy/chemistry/fingerprint.py
